Make ChainRulesCore and DensityInterface weak dependencies
